+++ HISTORY.rst
@@ -15,6 +15,7 @@ Next Release
   a snapshot report.
 * Fix some typos
 * Add history report view and connect it to `memote report history` call.
+* ``find_direct_metabolites`` detects and removes false positives.
 
 0.6.2 (2018-03-12)
 ------------------